Move interfaces to the new folder
