+++ frontend/src/components/annotation_interfaces/audio/audio_hotkeys.vue
