+++ frontend/src/components/annotation_interfaces/audio/audio_sidebar.vue
@@ -8,20 +8,7 @@
         `"
     >
         <v-expansion-panels multiple style="width: 350px;" accordion :value="open_panels">
-
-          <v-expansion-panel @change="on_change_expansion(0)">
-              <global_attributes_list
-                  v-if="global_attribute_groups_list && global_attribute_groups_list.length > 0"
-                :global_attribute_groups_list="global_attribute_groups_list"
-                :current_global_instance="current_global_instance"
-                :schema_id="schema_id"
-                :view_only_mode="false"
-                @attribute_change="attribute_change($event, true)"
-              ></global_attributes_list>
-          </v-expansion-panel>
-
-            <v-expansion-panel @change="on_change_expansion(1)" :disabled="!current_instance">
-
+            <!-- <v-expansion-panel @change="on_change_expansion(0)" :disabled="!current_instance">
                 <v-expansion-panel-header>
                     <strong>Attributes {{ !current_instance ? "(select instance)" : null }}</strong>
                 </v-expansion-panel-header>
@@ -38,8 +25,8 @@
                         key="attribute_groups_list"
                     />
                 </v-expansion-panel-content>
-            </v-expansion-panel>
-            <v-expansion-panel data-cy="instance-expansion-panel" @change="on_change_expansion(2)">
+            </v-expansion-panel> -->
+            <v-expansion-panel @change="on_change_expansion(1)">
                     <v-expansion-panel-header>
                         <strong>Instances</strong>
                     </v-expansion-panel-header>
@@ -56,7 +43,7 @@
                             <template v-slot:body="{ items }">
                                 <tbody v-if="items.length > 0 && !loading" style="cursor: pointer">
                                 <tr
-                                    v-for="(item, item_index) in items"
+                                    v-for="item in items"
                                     :key="item.id"
                                     @mouseover="on_hover_item(item)"
                                     @mouseleave="on_stop_hover_item"
@@ -68,19 +55,13 @@
                                     </td>
                                     <td class="centered-table-items">
                                             <v-icon 
-                                                v-if="item.type === 'relation'"
+                                                v-if="item.type === 'audio'"
                                                 :color="item.label_file.colour.hex"
                                             >
-                                                mdi-relation-one-to-one
-                                            </v-icon>
-                                            <v-icon 
-                                                v-if="item.type === 'text_token'"
-                                                :color="item.label_file.colour.hex"
-                                            >
-                                                mdi-label
+                                                mdi-waveform
                                             </v-icon>
                                     </td>
-                                    <td :data-cy="`label_name_${item_index}`" class="centered-table-items">
+                                    <td class="centered-table-items">
                                         {{ item.label_file.label.name }}
                                     </td>
                                     <td class="centered-table-items">
@@ -89,14 +70,12 @@
                                                     tooltip_message="Change Label Template"
                                                     icon="mdi-format-paint"
                                                     color="primary"
-                                                    :datacy="`change_label_${item_index}`"
                                                     :close_by_button="true"
                                                 >
                                                     <template slot="content">
                                                         <label_select_only
                                                             :label_file_list_prop="label_list"
                                                             :select_this_id_at_load="item.label_file_id"
-                                                            datacy="select_text_label"
                                                             @label_file="$emit('change_instance_label', { label: $event, instance: item })"
                                                         />
                                                     </template>
@@ -130,21 +109,18 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import instance_detail_list_view from "../annotation/instance_detail_list_view.vue";
-import button_with_menu from '../regular/button_with_menu.vue';
-import label_select_only from '../label/label_select_only.vue'
-import attribute_group_list from '../attribute/attribute_group_list.vue'
-import global_attributes_list from '../attribute/global_attributes_list.vue'
-
+import instance_detail_list_view from "../../annotation/instance_detail_list_view.vue";
+import button_with_menu from '../../regular/button_with_menu.vue';
+import label_select_only from '../../label/label_select_only.vue'
+import attribute_group_list from '../../attribute/attribute_group_list.vue'
 
 export default Vue.extend({
-    name: "text_sidepanel",
+    name: "audio_sidebar",
     components: {
         instance_detail_list_view,
         button_with_menu,
         label_select_only,
-        attribute_group_list,
-        global_attributes_list
+        attribute_group_list
     },
     props: {
         project_string_id: {
@@ -178,17 +154,11 @@ export default Vue.extend({
         loading: {
             type: Boolean,
             required: true
-        },
-        global_attribute_groups_list: {
-            type: Array,
-            default: null
-        },
-        current_global_instance: {
         }
     },
     data() {
         return {
-            open_panels: [0, 1]
+            open_panels: [0]
         }
     },
     computed: {
@@ -263,23 +233,23 @@ export default Vue.extend({
         },
         attribute_group_list_prop: function () {
 
-            if (!this.label_list
-                || !this.current_instance
-                || !this.per_instance_attribute_groups_list
-                || !this.current_instance.label_file) {
-                    return []
-            }
+        if (!this.label_list
+          || !this.current_instance
+          || !this.per_instance_attribute_groups_list
+          || !this.current_instance.label_file) {
+          return []
+        }
 
-            let attr_group_list = this.per_instance_attribute_groups_list.filter(elm =>{
-                let file_id_list = elm.label_file_list.map(label_file => label_file.id)
-                return file_id_list.includes(this.current_instance.label_file.id)
-            })
+        let attr_group_list = this.per_instance_attribute_groups_list.filter(elm =>{
+          let file_id_list = elm.label_file_list.map(label_file => label_file.id)
+          return file_id_list.includes(this.current_instance.label_file.id)
+        })
+        return attr_group_list;
 
-            return attr_group_list;
       },
-      attribute_change: function(event, is_global = false) {
-          this.$emit('on_update_attribute', event, is_global)
-      },
+      attribute_change: function(attribute) {
+          this.$emit('on_update_attribute', attribute)
+      }
     }
 })
 </script>
