+++ frontend/src/components/annotation_interfaces/audio/render_elements/waveform_selector.vue
