+++ frontend/src/components/annotation_interfaces/geo/geo_hotkeys.vue
