+++ frontend/src/components/annotation_interfaces/geo/geo_sidebar.vue
@@ -103,9 +103,9 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import instance_detail_list_view from "../annotation/instance_detail_list_view.vue";
-import button_with_menu from '../regular/button_with_menu.vue';
-import label_select_only from '../label/label_select_only.vue'
+import instance_detail_list_view from "../../annotation/instance_detail_list_view.vue";
+import button_with_menu from '../../regular/button_with_menu.vue';
+import label_select_only from '../../label/label_select_only.vue'
 export default Vue.extend({
     name: "geo_sidepanel",
     components: {
