+++ frontend/src/components/annotation_interfaces/geo/geo_toolbar.vue
@@ -414,9 +414,9 @@
 
 <script>
 import Vue from 'vue'
-import label_select_annotation from "../label/label_select_annotation.vue"
-import label_schema_selector from "../label/label_schema_selector.vue"
-import task_status from "../annotation/task_status.vue"
+import label_select_annotation from "../../label/label_select_annotation.vue"
+import label_schema_selector from "../../label/label_schema_selector.vue"
+import task_status from "../../annotation/task_status.vue"
 import geo_hotkeys from "./geo_hotkeys.vue"
 
 export default Vue.extend({
