+++ frontend/src/components/annotation_interfaces/text/render_elements/fast_label_menu.vue
