+++ frontend/src/components/annotation_interfaces/text/render_elements/selection.vue
