+++ frontend/src/components/annotation_interfaces/text/render_elements/text_context_menu.vue
