+++ frontend/src/components/annotation_interfaces/text/text_hotkeys.vue
