+++ frontend/src/components/annotation_interfaces/text/text_sidebar.vue
@@ -8,7 +8,20 @@
         `"
     >
         <v-expansion-panels multiple style="width: 350px;" accordion :value="open_panels">
-            <!-- <v-expansion-panel @change="on_change_expansion(0)" :disabled="!current_instance">
+
+          <v-expansion-panel @change="on_change_expansion(0)">
+              <global_attributes_list
+                  v-if="global_attribute_groups_list && global_attribute_groups_list.length > 0"
+                :global_attribute_groups_list="global_attribute_groups_list"
+                :current_global_instance="current_global_instance"
+                :schema_id="schema_id"
+                :view_only_mode="false"
+                @attribute_change="attribute_change($event, true)"
+              ></global_attributes_list>
+          </v-expansion-panel>
+
+            <v-expansion-panel @change="on_change_expansion(1)" :disabled="!current_instance">
+
                 <v-expansion-panel-header>
                     <strong>Attributes {{ !current_instance ? "(select instance)" : null }}</strong>
                 </v-expansion-panel-header>
@@ -25,8 +38,8 @@
                         key="attribute_groups_list"
                     />
                 </v-expansion-panel-content>
-            </v-expansion-panel> -->
-            <v-expansion-panel @change="on_change_expansion(1)">
+            </v-expansion-panel>
+            <v-expansion-panel data-cy="instance-expansion-panel" @change="on_change_expansion(2)">
                     <v-expansion-panel-header>
                         <strong>Instances</strong>
                     </v-expansion-panel-header>
@@ -43,7 +56,7 @@
                             <template v-slot:body="{ items }">
                                 <tbody v-if="items.length > 0 && !loading" style="cursor: pointer">
                                 <tr
-                                    v-for="item in items"
+                                    v-for="(item, item_index) in items"
                                     :key="item.id"
                                     @mouseover="on_hover_item(item)"
                                     @mouseleave="on_stop_hover_item"
@@ -55,13 +68,19 @@
                                     </td>
                                     <td class="centered-table-items">
                                             <v-icon 
-                                                v-if="item.type === 'audio'"
+                                                v-if="item.type === 'relation'"
                                                 :color="item.label_file.colour.hex"
                                             >
-                                                mdi-waveform
+                                                mdi-relation-one-to-one
+                                            </v-icon>
+                                            <v-icon 
+                                                v-if="item.type === 'text_token'"
+                                                :color="item.label_file.colour.hex"
+                                            >
+                                                mdi-label
                                             </v-icon>
                                     </td>
-                                    <td class="centered-table-items">
+                                    <td :data-cy="`label_name_${item_index}`" class="centered-table-items">
                                         {{ item.label_file.label.name }}
                                     </td>
                                     <td class="centered-table-items">
@@ -70,12 +89,14 @@
                                                     tooltip_message="Change Label Template"
                                                     icon="mdi-format-paint"
                                                     color="primary"
+                                                    :datacy="`change_label_${item_index}`"
                                                     :close_by_button="true"
                                                 >
                                                     <template slot="content">
                                                         <label_select_only
                                                             :label_file_list_prop="label_list"
                                                             :select_this_id_at_load="item.label_file_id"
+                                                            datacy="select_text_label"
                                                             @label_file="$emit('change_instance_label', { label: $event, instance: item })"
                                                         />
                                                     </template>
@@ -109,18 +130,21 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import instance_detail_list_view from "../annotation/instance_detail_list_view.vue";
-import button_with_menu from '../regular/button_with_menu.vue';
-import label_select_only from '../label/label_select_only.vue'
-import attribute_group_list from '../attribute/attribute_group_list.vue'
+import instance_detail_list_view from "../../annotation/instance_detail_list_view.vue";
+import button_with_menu from '../../regular/button_with_menu.vue';
+import label_select_only from '../../label/label_select_only.vue'
+import attribute_group_list from '../../attribute/attribute_group_list.vue'
+import global_attributes_list from '../../attribute/global_attributes_list.vue'
+
 
 export default Vue.extend({
-    name: "audio_sidebar",
+    name: "text_sidepanel",
     components: {
         instance_detail_list_view,
         button_with_menu,
         label_select_only,
-        attribute_group_list
+        attribute_group_list,
+        global_attributes_list
     },
     props: {
         project_string_id: {
@@ -154,11 +178,17 @@ export default Vue.extend({
         loading: {
             type: Boolean,
             required: true
+        },
+        global_attribute_groups_list: {
+            type: Array,
+            default: null
+        },
+        current_global_instance: {
         }
     },
     data() {
         return {
-            open_panels: [0]
+            open_panels: [0, 1]
         }
     },
     computed: {
@@ -233,23 +263,23 @@ export default Vue.extend({
         },
         attribute_group_list_prop: function () {
 
-        if (!this.label_list
-          || !this.current_instance
-          || !this.per_instance_attribute_groups_list
-          || !this.current_instance.label_file) {
-          return []
-        }
+            if (!this.label_list
+                || !this.current_instance
+                || !this.per_instance_attribute_groups_list
+                || !this.current_instance.label_file) {
+                    return []
+            }
 
-        let attr_group_list = this.per_instance_attribute_groups_list.filter(elm =>{
-          let file_id_list = elm.label_file_list.map(label_file => label_file.id)
-          return file_id_list.includes(this.current_instance.label_file.id)
-        })
-        return attr_group_list;
+            let attr_group_list = this.per_instance_attribute_groups_list.filter(elm =>{
+                let file_id_list = elm.label_file_list.map(label_file => label_file.id)
+                return file_id_list.includes(this.current_instance.label_file.id)
+            })
 
+            return attr_group_list;
       },
-      attribute_change: function(attribute) {
-          this.$emit('on_update_attribute', attribute)
-      }
+      attribute_change: function(event, is_global = false) {
+          this.$emit('on_update_attribute', event, is_global)
+      },
     }
 })
 </script>
