+++ frontend/src/components/annotation_interfaces/text/text_utils/closest_token.ts
