+++ frontend/src/components/annotation_interfaces/text/text_utils/draw_rects.ts
